Created 'shared' file and added utility functions.
