+++ screens/Contact/ContactDetailScreen.js
@@ -40,7 +40,7 @@ import KeyboardAccessory from 'react-native-sticky-keyboard-accessory';
 import ProgressBarAnimated from 'react-native-progress-bar-animated';
 import ModalFilterPicker from 'react-native-modal-filter-picker';
 import { Chip, Selectize } from 'react-native-material-selectize';
-
+import sharedTools from '../../shared';
 import KeyboardShift from '../../components/KeyboardShift';
 import {
   save,
@@ -180,182 +180,6 @@ const styles = StyleSheet.create({
     marginTop: 40,
   },
 });
-const diff = (obj1, obj2) => {
-  // Make sure an object to compare is provided
-  if (!obj2 || Object.prototype.toString.call(obj2) !== '[object Object]') {
-    return obj1;
-  }
-
-  //
-  // Variables
-  //
-
-  const diffs = {};
-  // let key;
-
-
-  //
-  // Methods
-  //
-
-  /**
-   * Check if two arrays are equal
-   * @param  {Array}   arr1 The first array
-   * @param  {Array}   arr2 The second array
-   * @return {Boolean}      If true, both arrays are equal
-   */
-  const arraysMatch = (value, other) => {
-    // Get the value type
-    const type = Object.prototype.toString.call(value);
-
-    // If the two objects are not the same type, return false
-    if (type !== Object.prototype.toString.call(other)) return false;
-
-    // If items are not an object or array, return false
-    if (['[object Array]', '[object Object]'].indexOf(type) < 0) return false;
-
-    // Compare the length of the length of the two items
-    const valueLen = type === '[object Array]' ? value.length : Object.keys(value).length;
-    const otherLen = type === '[object Array]' ? other.length : Object.keys(other).length;
-    if (valueLen !== otherLen) return false;
-
-    // Compare two items
-    const compare = (item1, item2) => {
-      // Get the object type
-      const itemType = Object.prototype.toString.call(item1);
-
-      // If an object or array, compare recursively
-      if (['[object Array]', '[object Object]'].indexOf(itemType) >= 0) {
-        if (!arraysMatch(item1, item2)) return false;
-      } else {
-        // Otherwise, do a simple comparison
-        // If the two items are not the same type, return false
-        if (itemType !== Object.prototype.toString.call(item2)) return false;
-
-        // Else if it's a function, convert to a string and compare
-        // Otherwise, just compare
-        if (itemType === '[object Function]') {
-          if (item1.toString() !== item2.toString()) return false;
-        } else if (item1 !== item2) return false;
-      }
-      return true;
-    };
-
-    // Compare properties
-    if (type === '[object Array]') {
-      for (let i = 0; i < valueLen; i++) {
-        if (compare(value[i], other[i]) === false) return false;
-      }
-    } else {
-      for (const key in value) {
-        if (Object.prototype.hasOwnProperty.call(value, key)) {
-          if (compare(value[key], other[key]) === false) return false;
-        }
-      }
-    }
-
-    // If nothing failed, return true
-    return true;
-  };
-
-  /**
-   * Compare two items and push non-matches to object
-   * @param  {*}      item1 The first item
-   * @param  {*}      item2 The second item
-   * @param  {String} key   The key in our object
-   */
-  const compare = (item1, item2, key) => {
-    // Get the object type
-    const type1 = Object.prototype.toString.call(item1);
-    const type2 = Object.prototype.toString.call(item2);
-
-    // If type2 is undefined it has been removed
-    if (type2 === '[object Undefined]') {
-      diffs[key] = null;
-      return;
-    }
-
-    // If items are different types
-    if (type1 !== type2) {
-      diffs[key] = item2;
-      return;
-    }
-
-    // If an object, compare recursively
-    if (type1 === '[object Object]') {
-      const objDiff = diff(item1, item2);
-      if (Object.keys(objDiff).length > 0) {
-        diffs[key] = objDiff;
-      }
-      return;
-    }
-
-    // If an array, compare
-    if (type1 === '[object Array]') {
-      if (!arraysMatch(item1, item2)) {
-        diffs[key] = item2;
-      }
-      return;
-    }
-
-    // Else if it's a function, convert to a string and compare
-    // Otherwise, just compare
-    if (type1 === '[object Function]') {
-      if (item1.toString() !== item2.toString()) {
-        diffs[key] = item2;
-      }
-    } else if (item1 !== item2) {
-      diffs[key] = item2;
-    }
-  };
-
-
-  //
-  // Compare our objects
-  //
-
-  // Loop through the first object
-  for (const key in obj1) {
-    if (Object.prototype.hasOwnProperty.call(obj1, key)) {
-      compare(obj1[key], obj2[key], key);
-    }
-  }
-
-  // Loop through the second object and find missing items
-  for (const key in obj2) {
-    if (Object.prototype.hasOwnProperty.call(obj2, key)) {
-      if (!obj1[key] && obj1[key] !== obj2[key]) {
-        diffs[key] = obj2[key];
-      }
-    }
-  }
-
-  // Return the object of differences
-  return diffs;
-};
-const formatDateToBackEnd = (dateString) => {
-  const dateObject = new Date(dateString);
-  const year = dateObject.getFullYear();
-  const month = (dateObject.getMonth() + 1) < 10 ? `0${dateObject.getMonth() + 1}` : (dateObject.getMonth() + 1);
-  const day = (dateObject.getDate()) < 10 ? `0${dateObject.getDate()}` : (dateObject.getDate());
-  const newDate = `${year}-${month}-${day}`;
-  return newDate;
-};
-const getOverallStatusSelectorColor = (contactStatus) => {
-  let newColor;
-  if (contactStatus === 'new' || contactStatus === 'unassigned' || contactStatus === 'closed') {
-    newColor = '#d9534f';
-  } else if (
-    contactStatus === 'unassignable'
-    || contactStatus === 'assigned'
-    || contactStatus === 'paused'
-  ) {
-    newColor = '#f0ad4e';
-  } else if (contactStatus === 'active') {
-    newColor = '#5cb85c';
-  }
-  return newColor;
-};
 
 class ContactDetailScreen extends React.Component {
   static navigationOptions = ({ navigation }) => {
@@ -526,7 +350,7 @@ class ContactDetailScreen extends React.Component {
         if (newState.contact.overall_status) {
           newState = {
             ...newState,
-            overallStatusBackgroundColor: getOverallStatusSelectorColor(newState.contact.overall_status),
+            overallStatusBackgroundColor: sharedTools.getSelectorColor(newState.contact.overall_status),
           };
         }
         if (prevState.contact.initial_comment) {
@@ -780,7 +604,7 @@ class ContactDetailScreen extends React.Component {
       contact: {
         ...unmodifiedContact,
       },
-      overallStatusBackgroundColor: getOverallStatusSelectorColor(unmodifiedContact.overall_status),
+      overallStatusBackgroundColor: sharedTools.getSelectorColor(unmodifiedContact.overall_status),
     }, () => {
       this.setCurrentTabIndex(currentTabIndex);
     });
@@ -881,7 +705,7 @@ class ContactDetailScreen extends React.Component {
         ...prevState.contact,
         overall_status: value,
       },
-      overallStatusBackgroundColor: getOverallStatusSelectorColor(value),
+      overallStatusBackgroundColor: sharedTools.getSelectorColor(value),
     }));
   };
 
@@ -902,7 +726,7 @@ class ContactDetailScreen extends React.Component {
     this.setState(prevState => ({
       contact: {
         ...prevState.contact,
-        baptism_date: formatDateToBackEnd(date),
+        baptism_date: sharedTools.formatDateToBackEnd(date),
       },
     }));
   };
@@ -912,7 +736,7 @@ class ContactDetailScreen extends React.Component {
     const { unmodifiedContact } = this.state;
     const contact = this.transformContactObject(this.state.contact, quickAction);
     let contactToSave = {
-      ...diff(unmodifiedContact, contact),
+      ...sharedTools.diff(unmodifiedContact, contact),
     };
     // Remove empty arrays
     Object.keys(contactToSave).forEach((key) => {
