+++ screens/Group/GroupDetailScreen.js
@@ -16,7 +16,7 @@ import {
   Platform,
 } from 'react-native';
 import PropTypes from 'prop-types';
-
+import sharedTools from '../../shared';
 import {
   Container,
   Content,
@@ -294,176 +294,6 @@ const styles = StyleSheet.create({
     marginTop: 40,
   },
 });
-const diff = (obj1, obj2) => {
-  // Make sure an object to compare is provided
-  if (!obj2 || Object.prototype.toString.call(obj2) !== '[object Object]') {
-    return obj1;
-  }
-
-  //
-  // Variables
-  //
-
-  const diffs = {};
-  // let key;
-
-
-  //
-  // Methods
-  //
-
-  /**
-   * Check if two arrays are equal
-   * @param  {Array}   arr1 The first array
-   * @param  {Array}   arr2 The second array
-   * @return {Boolean}      If true, both arrays are equal
-   */
-  const arraysMatch = (value, other) => {
-    // Get the value type
-    const type = Object.prototype.toString.call(value);
-
-    // If the two objects are not the same type, return false
-    if (type !== Object.prototype.toString.call(other)) return false;
-
-    // If items are not an object or array, return false
-    if (['[object Array]', '[object Object]'].indexOf(type) < 0) return false;
-
-    // Compare the length of the length of the two items
-    const valueLen = type === '[object Array]' ? value.length : Object.keys(value).length;
-    const otherLen = type === '[object Array]' ? other.length : Object.keys(other).length;
-    if (valueLen !== otherLen) return false;
-
-    // Compare two items
-    const compare = (item1, item2) => {
-      // Get the object type
-      const itemType = Object.prototype.toString.call(item1);
-
-      // If an object or array, compare recursively
-      if (['[object Array]', '[object Object]'].indexOf(itemType) >= 0) {
-        if (!arraysMatch(item1, item2)) return false;
-      } else {
-        // Otherwise, do a simple comparison
-        // If the two items are not the same type, return false
-        if (itemType !== Object.prototype.toString.call(item2)) return false;
-
-        // Else if it's a function, convert to a string and compare
-        // Otherwise, just compare
-        if (itemType === '[object Function]') {
-          if (item1.toString() !== item2.toString()) return false;
-        } else if (item1 !== item2) return false;
-      }
-      return true;
-    };
-
-    // Compare properties
-    if (type === '[object Array]') {
-      for (let i = 0; i < valueLen; i++) {
-        if (compare(value[i], other[i]) === false) return false;
-      }
-    } else {
-      for (const key in value) {
-        if (Object.prototype.hasOwnProperty.call(value, key)) {
-          if (compare(value[key], other[key]) === false) return false;
-        }
-      }
-    }
-
-    // If nothing failed, return true
-    return true;
-  };
-
-  /**
-   * Compare two items and push non-matches to object
-   * @param  {*}      item1 The first item
-   * @param  {*}      item2 The second item
-   * @param  {String} key   The key in our object
-   */
-  const compare = (item1, item2, key) => {
-    // Get the object type
-    const type1 = Object.prototype.toString.call(item1);
-    const type2 = Object.prototype.toString.call(item2);
-
-    // If type2 is undefined it has been removed
-    if (type2 === '[object Undefined]') {
-      diffs[key] = null;
-      return;
-    }
-
-    // If items are different types
-    if (type1 !== type2) {
-      diffs[key] = item2;
-      return;
-    }
-
-    // If an object, compare recursively
-    if (type1 === '[object Object]') {
-      const objDiff = diff(item1, item2);
-      if (Object.keys(objDiff).length > 0) {
-        diffs[key] = objDiff;
-      }
-      return;
-    }
-
-    // If an array, compare
-    if (type1 === '[object Array]') {
-      if (!arraysMatch(item1, item2)) {
-        diffs[key] = item2;
-      }
-      return;
-    }
-
-    // Else if it's a function, convert to a string and compare
-    // Otherwise, just compare
-    if (type1 === '[object Function]') {
-      if (item1.toString() !== item2.toString()) {
-        diffs[key] = item2;
-      }
-    } else if (item1 !== item2) {
-      diffs[key] = item2;
-    }
-  };
-
-
-  //
-  // Compare our objects
-  //
-
-  // Loop through the first object
-  for (const key in obj1) {
-    if (Object.prototype.hasOwnProperty.call(obj1, key)) {
-      compare(obj1[key], obj2[key], key);
-    }
-  }
-
-  // Loop through the second object and find missing items
-  for (const key in obj2) {
-    if (Object.prototype.hasOwnProperty.call(obj2, key)) {
-      if (!obj1[key] && obj1[key] !== obj2[key]) {
-        diffs[key] = obj2[key];
-      }
-    }
-  }
-
-  // Return the object of differences
-  return diffs;
-};
-const formatDateToBackEnd = (dateString) => {
-  const dateObject = new Date(dateString);
-  const year = dateObject.getFullYear();
-  const month = (dateObject.getMonth() + 1) < 10 ? `0${dateObject.getMonth() + 1}` : (dateObject.getMonth() + 1);
-  const day = (dateObject.getDate()) < 10 ? `0${dateObject.getDate()}` : (dateObject.getDate());
-  const newDate = `${year}-${month}-${day}`;
-  return newDate;
-};
-const getStatusSelectorColor = (groupStatus) => {
-  let newColor;
-  if (groupStatus === 'inactive') {
-    newColor = '#d9534f';
-  } else if (groupStatus === 'active') {
-    newColor = '#5cb85c';
-  }
-  return newColor;
-};
 
 const initialState = {
   group: {},
@@ -624,7 +454,7 @@ class GroupDetailScreen extends React.Component {
         if (newState.group.group_status) {
           newState = {
             ...newState,
-            groupStatusBackgroundColor: getStatusSelectorColor(newState.group.group_status),
+            groupStatusBackgroundColor: sharedTools.getSelectorColor(newState.group.group_status),
           };
         }
       }
@@ -904,7 +734,7 @@ class GroupDetailScreen extends React.Component {
       group: {
         ...unmodifiedGroup,
       },
-      groupStatusBackgroundColor: getStatusSelectorColor(unmodifiedGroup.group_status),
+      groupStatusBackgroundColor: sharedTools.getSelectorColor(unmodifiedGroup.group_status),
     }, () => {
       this.setCurrentTabIndex(currentTabIndex);
     });
@@ -949,7 +779,7 @@ class GroupDetailScreen extends React.Component {
     this.setState(prevState => ({
       group: {
         ...prevState.group,
-        start_date: formatDateToBackEnd(value),
+        start_date: sharedTools.formatDateToBackEnd(value),
       },
     }));
   };
@@ -958,7 +788,7 @@ class GroupDetailScreen extends React.Component {
     this.setState(prevState => ({
       group: {
         ...prevState.group,
-        end_date: formatDateToBackEnd(value),
+        end_date: sharedTools.formatDateToBackEnd(value),
       },
     }));
   };
@@ -1249,7 +1079,7 @@ class GroupDetailScreen extends React.Component {
     const { unmodifiedGroup } = this.state;
     const group = this.transformGroupObject(this.state.group);
     let groupToSave = {
-      ...diff(unmodifiedGroup, group),
+      ...sharedTools.diff(unmodifiedGroup, group),
     };
     if (this.state.group.title) {
       groupToSave = {
